Move TaakUser to src/config
