+++ backend/src/main/java/com/team8/taaks/config/TaakUserDetailManager.java
@@ -1,4 +1,4 @@
-package com.team8.taaks.model;
+package com.team8.taaks.config;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -7,6 +7,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.UserDetailsManager;
 
+import com.team8.taaks.model.TaakUser;
 import com.team8.taaks.repository.TaakUserRepository;
 
 public class TaakUserDetailManager implements UserDetailsManager{
